requirements.txt
 fastapi 
 uvicorn[standard] 
 sqlalchemy 
 psycopg2-binary 
 python-jose[cryptography] 
 passlib[bcrypt] 
 python-multipart 
 pydantic-settings 
 googlemaps 
 python-dotenv 
 websockets